rename inviz to bsavi
