+++ app/minimal_example.py
@@ -1,6 +1,6 @@
 import pandas as pd
 import numpy as np
-import inviz as iv
+import bsavi as bsv
 import holoviews as hv
 from holoviews import opts
 
@@ -35,6 +35,6 @@ sin_latex = {'x': 'x', 'y': r'\alpha \sin{\beta x} + \gamma \sin{\delta x}'}
 poly_opts = opts.Curve(framewise=True)
 sin_opts = opts.Curve(framewise=True)
 
-polynomials = iv.Observable(name='polynomials', parameters=[poly], plot_type=['Curve'], plot_opts=poly_opts, latex_labels=poly_latex)
-sine_wave = iv.Observable(name='Composite Sine Wave', parameters=[sin], plot_type=['Curve'], plot_opts=sin_opts, latex_labels=sin_latex)
-iv.viz(like, observables=[polynomials, sine_wave], latex_dict=like_latex).servable()
+polynomials = bsv.Observable(name='polynomials', parameters=[poly], plot_type=['Curve'], plot_opts=poly_opts, latex_labels=poly_latex)
+sine_wave = bsv.Observable(name='Composite Sine Wave', parameters=[sin], plot_type=['Curve'], plot_opts=sin_opts, latex_labels=sin_latex)
+bsv.viz(like, observables=[polynomials, sine_wave], latex_dict=like_latex).servable()
