+++ app/planck_mnu.py
@@ -3,11 +3,11 @@ mycosmo = pd.read_json('../data/planck2018/plikHM_TTTEEE_lowl_lowE_lensing/power
 chains = mycosmo.drop(columns=['p(k)', 'cl_tt', 'cl_ee'])
 class_results = mycosmo[['p(k)', 'cl_tt', 'cl_ee']]
 
-from inviz.cosmo import load_params
+from bsavi.cosmo import load_params
 
 params_with_latex = load_params('../data/planck2018/plikHM_TTTEEE_lowl_lowE_lensing/base_mnu_plikHM_TTTEEE_lowl_lowE_lensing.paramnames')
 
-import inviz as iv
+import bsavi as bsv
 import holoviews as hv
 from holoviews import opts
 
@@ -21,7 +21,7 @@ ps_latex = {
     'Cl_ee': '[{\ell(\ell+1)}/{2\pi}]~C_{\ell}^{EE}',
 }
 
-power_spectra = iv.Observable(
+power_spectra = bsv.Observable(
     name=['P(k)', 'Lensed Cl_TT', 'Lensed Cl_EE'], 
     parameters=class_results,
     plot_type='Curve',
@@ -29,4 +29,4 @@ power_spectra = iv.Observable(
     latex_labels=ps_latex
 )
 
-iv.viz(chains, [power_spectra], latex_dict=params_with_latex).servable()
+bsv.viz(chains, [power_spectra], latex_dict=params_with_latex).servable()
