+++ docs/source/class_examples.rst
@@ -5,14 +5,14 @@ The `CLASS code <https://lesgourg.github.io/class_public/class.html>`_ is widely
 by cosmologists to simulate the evolution of linear perturbations in the universe and
 to compute CMB and large scale structure observables. CLASS accepts cosmological parameters 
 as input, which we will get from the Planck 2018 results. The following code demonstrates 
-using InViz to visualize the Planck 2018 results in parameter space and their relationship 
+using BSAVI to visualize the Planck 2018 results in parameter space and their relationship 
 with three observables: the matter power spectrum :math:`P(k)`, and the power spectra of 
 the CMB temperature :math:`C_{l}^{TT}` and polarization :math:`C_{l}^{EE}`.
 
 We will look at Table 6.7 from the 
 `Planck 2018 table of parameters <https://wiki.cosmos.esa.int/planck-legacy-archive/images/4/43/Baseline_params_table_2018_68pc_v2.pdf>`_.
 The data required for this tutorial is located 
-`here <https://github.com/wen-jams/inviz/tree/main/data/planck2018/plikHM_TTTEEE_lowl_lowE_lensing>`_.
+`here <https://github.com/wen-jams/bsavi/tree/main/data/planck2018/plikHM_TTTEEE_lowl_lowE_lensing>`_.
 
 
 Pre-Computed Observables
@@ -34,7 +34,7 @@ plain text and LaTeX. ``load_params`` will return a dictionary of these labels w
 
 .. code-block:: python
 
-    from inviz.cosmo import load_params
+    from bsavi.cosmo import load_params
 
     params_with_latex = load_params('../data/planck2018/plikHM_TTTEEE_lowl_lowE_lensing/base_mnu_plikHM_TTTEEE_lowl_lowE_lensing.paramnames')
 
@@ -42,7 +42,7 @@ Now we are ready to construct an Observable which combines the observable data w
 
 .. code-block:: python
 
-    import inviz as iv
+    import bsavi as bsv
     from holoviews import opts
 
     curve_opts = opts.Curve(logx=True)
@@ -55,7 +55,7 @@ Now we are ready to construct an Observable which combines the observable data w
         'Cl_ee': '[{\ell(\ell+1)}/{2\pi}]~C_{\ell}^{EE}',
     }
 
-    power_spectra = iv.Observable(
+    power_spectra = bsv.Observable(
         name=['P(k)', 'Lensed Cl_TT', 'Lensed Cl_EE'], 
         parameters=class_results,
         plot_type='Curve',
@@ -66,7 +66,7 @@ Finally, produce the visualization with:
 
 .. code-block:: python
 
-    iv.viz(data=chains, observables=[power_spectra], latex_dict=params_with_latex).servable()
+    bsv.viz(data=chains, observables=[power_spectra], latex_dict=params_with_latex).servable()
 
 .. image:: ../../images/bsavi-mainpage.gif
 
@@ -81,8 +81,8 @@ This next example requires that you have `Classy, the Python wrapper for CLASS
 
     # imports
     import pandas as pd
-    import inviz as iv
-    from inviz import cosmo
+    import bsavi as bsv
+    from bsavi import cosmo
     from holoviews import opts
 
 As before, we load in the ``.paramnames`` file to get a dict with all the parameter names and their LaTeX code.
@@ -135,7 +135,7 @@ Now, we'll go through the same steps as above to set up our Observable.
         'Cl_ee': '[{\ell(\ell+1)}/{2\pi}]~C_{\ell}^{EE}',
     }
 
-    power_spectra = iv.Observable(
+    power_spectra = bsv.Observable(
         name=['P(k)', 'Cl_TT', 'Cl_EE'], 
         myfunc=cosmo.run_class,
         myfunc_args=(inclass,),
@@ -152,7 +152,7 @@ And finally, we can produce the visualization with
 
 .. code-block:: python
 
-    iv.viz(data=chains, observables=[power_spectra], latex_dict=params_with_latex).servable()
+    bsv.viz(data=chains, observables=[power_spectra], latex_dict=params_with_latex).servable()
 
 
 .. image:: ../../images/bsavi-mainpage.gif
