+++ docs/source/conf.py
@@ -6,7 +6,7 @@
 # -- Project information -----------------------------------------------------
 # https://www.sphinx-doc.org/en/master/usage/configuration.html#project-information
 
-project = 'InViz'
+project = 'BSAVI'
 copyright = '2023, James Wen'
 author = 'James Wen'
 release = '0.3.7'
