+++ src/bsavi/cosmo/__init__.py
@@ -1,0 +1,1 @@
+from .cosmo import *
