+++ src/inviz.egg-info/SOURCES.txt
@@ -2,12 +2,12 @@ LICENSE
 MANIFEST.in
 README.md
 setup.py
-src/inviz/__init__.py
-src/inviz/inviz.py
-src/inviz.egg-info/PKG-INFO
-src/inviz.egg-info/SOURCES.txt
-src/inviz.egg-info/dependency_links.txt
-src/inviz.egg-info/requires.txt
-src/inviz.egg-info/top_level.txt
-src/inviz/cosmo/__init__.py
-src/inviz/cosmo/cosmo.py
+src/bsavi/__init__.py
+src/bsavi/bsavi.py
+src/bsavi.egg-info/PKG-INFO
+src/bsavi.egg-info/SOURCES.txt
+src/bsavi.egg-info/dependency_links.txt
+src/bsavi.egg-info/requires.txt
+src/bsavi.egg-info/top_level.txt
+src/bsavi/cosmo/__init__.py
+src/bsavi/cosmo/cosmo.py
