--- src/inviz/__init__.py
@@ -1,1 +1,0 @@
-from .inviz import *
