--- src/inviz/cosmo/__init__.py
@@ -1,1 +1,0 @@
-from .cosmo import *
